Refactor all memory impls into the memory folder
