+++ synthesizer/src/store/helpers/memory/mod.rs
@@ -16,3 +16,18 @@
 
 pub mod internal;
 pub use internal::*;
+
+mod block;
+pub use block::*;
+
+mod consensus;
+pub use consensus::*;
+
+mod program;
+pub use program::*;
+
+mod transaction;
+pub use transaction::*;
+
+mod transition;
+pub use transition::*;
