+++ synthesizer/src/store/transaction/deployment.rs
@@ -23,8 +23,7 @@ use crate::{
     program::Program,
     snark::{Certificate, Proof, VerifyingKey},
     store::{
-        helpers::{memory::MemoryMap, Map, MapRead},
-        TransitionMemory,
+        helpers::{Map, MapRead},
         TransitionStorage,
         TransitionStore,
     },
@@ -456,112 +455,6 @@ pub trait DeploymentStorage<N: Network>: Clone + Send + Sync {
     }
 }
 
-/// An in-memory deployment storage.
-#[derive(Clone)]
-#[allow(clippy::type_complexity)]
-pub struct DeploymentMemory<N: Network> {
-    /// The ID map.
-    id_map: MemoryMap<N::TransactionID, ProgramID<N>>,
-    /// The edition map.
-    edition_map: MemoryMap<ProgramID<N>, u16>,
-    /// The reverse ID map.
-    reverse_id_map: MemoryMap<(ProgramID<N>, u16), N::TransactionID>,
-    /// The owner map.
-    owner_map: MemoryMap<(ProgramID<N>, u16), ProgramOwner<N>>,
-    /// The program map.
-    program_map: MemoryMap<(ProgramID<N>, u16), Program<N>>,
-    /// The verifying key map.
-    verifying_key_map: MemoryMap<(ProgramID<N>, Identifier<N>, u16), VerifyingKey<N>>,
-    /// The certificate map.
-    certificate_map: MemoryMap<(ProgramID<N>, Identifier<N>, u16), Certificate<N>>,
-    /// The fee map.
-    fee_map: MemoryMap<N::TransactionID, (N::TransitionID, N::StateRoot, Option<Proof<N>>)>,
-    /// The reverse fee map.
-    reverse_fee_map: MemoryMap<N::TransitionID, N::TransactionID>,
-    /// The transition store.
-    transition_store: TransitionStore<N, TransitionMemory<N>>,
-}
-
-#[rustfmt::skip]
-impl<N: Network> DeploymentStorage<N> for DeploymentMemory<N> {
-    type IDMap = MemoryMap<N::TransactionID, ProgramID<N>>;
-    type EditionMap = MemoryMap<ProgramID<N>, u16>;
-    type ReverseIDMap = MemoryMap<(ProgramID<N>, u16), N::TransactionID>;
-    type OwnerMap = MemoryMap<(ProgramID<N>, u16), ProgramOwner<N>>;
-    type ProgramMap = MemoryMap<(ProgramID<N>, u16), Program<N>>;
-    type VerifyingKeyMap = MemoryMap<(ProgramID<N>, Identifier<N>, u16), VerifyingKey<N>>;
-    type CertificateMap = MemoryMap<(ProgramID<N>, Identifier<N>, u16), Certificate<N>>;
-    type FeeMap = MemoryMap<N::TransactionID, (N::TransitionID, N::StateRoot, Option<Proof<N>>)>;
-    type ReverseFeeMap = MemoryMap<N::TransitionID, N::TransactionID>;
-    type TransitionStorage = TransitionMemory<N>;
-
-    /// Initializes the deployment storage.
-    fn open(transition_store: TransitionStore<N, Self::TransitionStorage>) -> Result<Self> {
-        Ok(Self {
-            id_map: MemoryMap::default(),
-            edition_map: MemoryMap::default(),
-            reverse_id_map: MemoryMap::default(),
-            owner_map: MemoryMap::default(),
-            program_map: MemoryMap::default(),
-            verifying_key_map: MemoryMap::default(),
-            certificate_map: MemoryMap::default(),
-            fee_map: MemoryMap::default(),
-            reverse_fee_map: MemoryMap::default(),
-            transition_store,
-        })
-    }
-
-    /// Returns the ID map.
-    fn id_map(&self) -> &Self::IDMap {
-        &self.id_map
-    }
-
-    /// Returns the edition map.
-    fn edition_map(&self) -> &Self::EditionMap {
-        &self.edition_map
-    }
-
-    /// Returns the reverse ID map.
-    fn reverse_id_map(&self) -> &Self::ReverseIDMap {
-        &self.reverse_id_map
-    }
-
-    /// Returns the owner map.
-    fn owner_map(&self) -> &Self::OwnerMap {
-        &self.owner_map
-    }
-
-    /// Returns the program map.
-    fn program_map(&self) -> &Self::ProgramMap {
-        &self.program_map
-    }
-
-    /// Returns the verifying key map.
-    fn verifying_key_map(&self) -> &Self::VerifyingKeyMap {
-        &self.verifying_key_map
-    }
-
-    /// Returns the certificate map.
-    fn certificate_map(&self) -> &Self::CertificateMap {
-        &self.certificate_map
-    }
-
-    /// Returns the fee map.
-    fn fee_map(&self) -> &Self::FeeMap {
-        &self.fee_map
-    }
-
-    /// Returns the reverse fee map.
-    fn reverse_fee_map(&self) -> &Self::ReverseFeeMap {
-        &self.reverse_fee_map
-    }
-
-    /// Returns the transition store.
-    fn transition_store(&self) -> &TransitionStore<N, Self::TransitionStorage> {
-        &self.transition_store
-    }
-}
-
 /// The deployment store.
 #[derive(Clone)]
 pub struct DeploymentStore<N: Network, D: DeploymentStorage<N>> {
@@ -733,6 +626,7 @@ impl<N: Network, D: DeploymentStorage<N>> DeploymentStore<N, D> {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::store::helpers::memory::DeploymentMemory;
 
     #[test]
     fn test_insert_get_remove() {
